+++ apps/server/prisma/migrations/20260315163817_add_project_member/migration.sql
@@ -1,0 +1,18 @@
+-- CreateTable
+CREATE TABLE "ProjectMember" (
+    "id" TEXT NOT NULL,
+    "projectId" TEXT NOT NULL,
+    "memberProjectId" TEXT NOT NULL,
+    "createdAt" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
+
+    CONSTRAINT "ProjectMember_pkey" PRIMARY KEY ("id")
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "ProjectMember_projectId_memberProjectId_key" ON "ProjectMember"("projectId", "memberProjectId");
+
+-- AddForeignKey
+ALTER TABLE "ProjectMember" ADD CONSTRAINT "ProjectMember_projectId_fkey" FOREIGN KEY ("projectId") REFERENCES "Project"("id") ON DELETE RESTRICT ON UPDATE CASCADE;
+
+-- AddForeignKey
+ALTER TABLE "ProjectMember" ADD CONSTRAINT "ProjectMember_memberProjectId_fkey" FOREIGN KEY ("memberProjectId") REFERENCES "Project"("id") ON DELETE RESTRICT ON UPDATE CASCADE;